Use proxy wiring for consensus engine
